Implemented basic database
In the process of testing the basic database, it was necessary to
revamp the majority of the project structure. The database its
self is currently incredibly immature, it contains no niceties and
was directly copied from a previous implementation I built. While
it is a good basis, it needs a lot of work to be functional.