Set up templating
The templating engine works and can now be used. Bootstrap and its
dependencies are imported from a CDN and so should work and update
as bootstrap leaves alpha. We'll need to swap it out for live at
some point in the future.

I also fixed some minor errors elsewhere, namely that running the
application from the command line wouldn't actually work because of
some incorrect setup on my part.