Login and Register pages now inform users of errors
This is mainly a quality of life change, the forms now return errors
to the user on the frontend so that they can be informed of how exactly
their registration or login process has failed. There are still
improvements to be made here, but this is a huge update in terms of raw
usability.